move public things to public domain, refactor policies
