--- app/policies/challenges/test_set_policy.rb
@@ -1,33 +1,0 @@
-module Challenges
-  class TestSetPolicy < ApplicationPolicy
-    class Scope < ApplicationPolicy::Scope
-      def resolve
-        Current.challenge ? Current.challenge.test_sets : TestSet.none
-      end
-    end
-
-    def show?
-      admin_or_challenge_editor?
-    end
-
-    def new?
-      admin_or_challenge_editor?
-    end
-
-    def create?
-      admin_or_challenge_editor?
-    end
-
-    def edit?
-      admin_or_challenge_editor?
-    end
-
-    def update?
-      admin_or_challenge_editor?
-    end
-
-    def destroy?
-      admin_or_challenge_editor?
-    end
-  end
-end
