--- app/policies/challenges/test_sets/leaderboards_policy.rb
@@ -1,11 +1,0 @@
-module Challenges
-  module TestSets
-    class LeaderboardsPolicy < ApplicationPolicy
-      class Scope < ApplicationPolicy::Scope
-        def resolve
-          leaderboard_released? ? Current.challenge.test_sets : TestSet.none
-        end
-      end
-    end
-  end
-end
