+++ app/views/challenges/public/models/_model.html.erb
