+++ app/views/challenges/public/models/_scores.html.erb
