+++ app/views/challenges/public/models/index.html.erb
