+++ app/views/challenges/public/models/show.html.erb
