+++ app/views/challenges/public/tasks/_menu.html.erb
