+++ app/views/challenges/public/tasks/_task.html.erb
