+++ app/views/challenges/public/tasks/index.html.erb
