+++ app/views/challenges/public/tasks/leaderboards/_filters.html.erb
