+++ app/views/challenges/public/tasks/leaderboards/_leaderboard.html.erb
