+++ app/views/challenges/public/tasks/show.html.erb
@@ -1,22 +1,13 @@
 <% content_for :head do %>
   <%= turbo_refreshes_with method: :morph, scroll: :preserve %>
 <% end %>
-
 <% breadcrumb :task, @task %>
-
 <%= mltop_title @task.name %>
-
 <%= render "menu", task: @task %>
-
 <%= @task.description %>
-
 <% @task_test_sets.each do |task_test_set| %>
   <h2 class="text-2xl font-bold tracking-tight sm:text-4xl mb-8 mt-8 text-center">
     <%= link_to "#{task_test_set.test_set.name} test set", test_set_path(task_test_set.test_set) %>
   </h2>
-
-  <%= render "challenges/test_sets/show", test_set: task_test_set.test_set, task_test_sets: [task_test_set] %>
+  <%= render "challenges/public/test_sets/show", test_set: task_test_set.test_set, task_test_sets: [task_test_set] %>
 <% end %>
-
-
-
