+++ app/views/challenges/public/test_sets/_menu.html.erb
