+++ app/views/challenges/public/test_sets/leaderboards/_leaderboard.html.erb
