+++ app/views/challenges/public/test_sets/show.html.erb
@@ -1,11 +1,7 @@
 <% content_for :head do %>
   <%= turbo_refreshes_with method: :morph, scroll: :preserve %>
 <% end %>
-
 <% breadcrumb :test_set, @test_set %>
-
 <%= mltop_title @test_set.name %>
-
-<%= render "challenges/test_sets/menu", test_set: @test_set %>
-
+<%= render "challenges/public/test_sets/menu", test_set: @test_set %>
 <%= render "show", test_set: @test_set, task_test_sets: @task_test_sets %>
